feat(utils): extract JSONPlaceholder clients into flat modules
